fix(react): restore rating browser test categorization
